Exploding cards
A card game has been made. At the start, the cards are shuffled randomly (you can see it on the console), you draw cards and if you touch a bomb card, a message appears saying that you have lost. When you lose, the button changes and says restart game and starts again.